refactor: UnitOfWork / Repositories 型を domain 層に移動する (#733)
application/common/unit-of-work.ts を domain/common/unit-of-work.ts に移動し、
DDD の依存方向（infrastructure → domain のみ）を正す。
これにより infrastructure → application の ESLint レイヤー境界例外を 2 件除去。
