+++ server/application/circle/circle-membership-service.ts
@@ -5,7 +5,7 @@ import type { createAccessService } from "@/server/application/authz/access-serv
 import type {
   Repositories,
   UnitOfWork,
-} from "@/server/application/common/unit-of-work";
+} from "@/server/domain/common/unit-of-work";
 import {
   assertCanAddCircleMemberWithRole,
   assertCanChangeCircleMemberRole,
