+++ server/infrastructure/repository/in-memory/in-memory-unit-of-work.ts
@@ -1,7 +1,7 @@
 import type {
   Repositories,
   UnitOfWork,
-} from "@/server/application/common/unit-of-work";
+} from "@/server/domain/common/unit-of-work";
 
 export const createInMemoryUnitOfWork = (repos: Repositories): UnitOfWork => {
   return async <T>(operation: (repos: Repositories) => Promise<T>) =>
